Added the use of the ButtonCommands Prettied Code Added Font, Scrollbars
